+++ packages/jest-dom-mocks/CHANGELOG.md
@@ -10,9 +10,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
 ## Unreleased
 
-## Added
+## [2.8.0] - 2019-10-03
 
-- Added the `Connection` class for to allow mocking `navigator.connection` in tests
+- Added the `Connection` class for to allow mocking `navigator.connection` in tests [#1083](https://github.com/Shopify/quilt/pull/1083/files)
 
 ## [2.7.1] - 2019-07-03